+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod addr;
 pub mod archive;
 pub mod opt;
 mod prelude;
+pub mod raw;
 pub mod timeout;
 pub mod tools;
 pub mod types;